Merging all dry run additions in to master

+++ maestro/alias.py
@@ -12,6 +12,18 @@ iam = boto3.resource('iam')
 roles = boto3.client('iam')
 DOC = ARGS.filename
 
+class color:
+   PURPLE = '\033[95m'
+   CYAN = '\033[96m'
+   DARKCYAN = '\033[36m'
+   BLUE = '\033[94m'
+   GREEN = '\033[92m'
+   YELLOW = '\033[93m'
+   RED = '\033[91m'
+   BOLD = '\033[1m'
+   UNDERLINE = '\033[4m'
+   END = '\033[0m'
+
 def json_parser():
   with open('%s' % DOC) as json_data:
     read = json.load(json_data)
@@ -34,17 +46,45 @@ def alias_creation():
   for names in load['Aliases']:    
     aliases.append(names['Name'])
 
-  if ARGS.alias in aliases:
-    print("Attempting to update alias %s" % ARGS.alias)
-    if alias_update():
-      return True
+  if ARGS.alias:    
+    if ARGS.alias in aliases:
+      if ARGS.publish:
+        if alias_update():
+          return True
+      else:
+        ask = input("Alias exists, would you like to update it (y/n)? ")
+        if ask == 'y':
+          if alias_update():
+            print("Attempting to update alias %s" % ARGS.alias)
+            return True
+          else:
+            return False
+        else:
+          print("Exiting!")
+          sys.exit(1)
+    else:
+      if ARGS.dry_run:
+        alias_name = ARGS.alias
+        print(color.PURPLE + "***Dry run option enabled***" + color.END)
+        print(color.PURPLE + "Would have created alias: %s for %s" % (ARGS.alias, lambda_name) + color.END)
+      else:
+        alias_name = ARGS.alias
+        pass
   else:
-    pass
+    get_alias_name = input("What would you like this alias to be called? ")
+    alias_name = get_alias_name.lower()
+    if alias_name in aliases:
+      ask = input("Alias exists, would you like to update it (y/n)? ")
+      if ask == 'y':
+        if alias_update():
+          print("Attempting to update alias %s" % ARGS.alias)
+          return True
+        else:
+          return False
+      else:
+        print("Exiting!")
+        sys.exit(1)
 
-  #get_alias_name = input("What would you like this alias to be called? ")
-  #alias_name = get_alias_name.lower()
-
-  alias_name = ARGS.alias
 
   versions = client.list_versions_by_function(
                     FunctionName='%s' % lambda_name,
@@ -60,8 +100,17 @@ def alias_creation():
       print('version: %s' % version['Version'])
       avail_versions.append(version['Version'])
 
-  function_version = input("What version would you like to create an alias for? ")
+  if ARGS.publish:
+    function_version = max(avail_versions)
+  else:
+    function_version = input("What version would you like to create an alias for? ")
   
+  if ARGS.dry_run:
+    print(color.PURPLE + "Dry run creating alias '%s' for version '%s' on lambda '%s'" % (alias_name, function_version, lambda_name) + color.END)
+    return True
+  else:
+    pass
+
   if function_version in avail_versions: 
     try:
       add_alias = client.create_alias(
@@ -96,18 +145,26 @@ def alias_destroy():
     aliases.append(names['Name'])
 
   if len(aliases) != 0:
-    alias_name = input("What alias would you like to delete? ")
+    if ARGS.alias:
+      alias_name = ARGS.alias
+    else:
+      alias_name = input("What alias would you like to delete? ")
 
     if alias_name in aliases:
-      try:
-        delete_alias = client.delete_alias(
-            FunctionName='%s' % lambda_name,
-            Name='%s' % alias_name
-        )
-        print("Alias successfully deleted")
+      if ARGS.dry_run:
+        print(color.PURPLE + "***Dry run option enabled***" + color.END)
+        print(color.PURPLE + "Dry run deleting alias %s on %s" % (alias_name, lambda_name) + color.END)
         return True
-      except ClientError as error:
-        print(error.response['Error']['Message'])
+      else:
+        try:
+          delete_alias = client.delete_alias(
+              FunctionName='%s' % lambda_name,
+              Name='%s' % alias_name
+          )
+          print("Alias successfully deleted")
+          return True
+        except ClientError as error:
+          print(error.response['Error']['Message'])
     else:
       print("That alias does not exist, please check the list of existing aliases and try again")
   else:
@@ -138,33 +195,41 @@ def alias_update():
     print("Function Version: '%s' has alias: '%s'" % (names['FunctionVersion'], names['Name']))     
     aliases.append(names['Name'])
   print("\n")
-  #print("Available Versions")
   for version in versions:
     if version['Version'] != 0:
-      #print('version: %s' % version['Version'])
       avail_versions.append(version['Version'])
 
   if len(aliases) != 0:
-    #alias_name = input("What alias would you like to update? ")
-    alias_name = ARGS.alias
-    #version_update = input("What version would you like to assign the update alias to? ")
-    version_update = max(avail_versions)
+    if ARGS.alias:
+      alias_name = ARGS.alias
+    else:
+      alias_name = input("What alias would you like to update? ")
+
+    if ARGS.publish:
+      version_update = max(avail_versions)
+    else:
+      version_update = input("What version would you like to assign the update alias to? ")
 
     if alias_name in aliases:
       if version_update in avail_versions:
-        try:
-          update_alias = client.update_alias(
-                          FunctionName='%s' % lambda_name,
-                          Name='%s' % alias_name,
-                          FunctionVersion='%s' % version_update,
-                        )
-          if update_alias['ResponseMetadata']['HTTPStatusCode'] == 200:
-            print("Lamda '%s' version '%s' alias '%s' has been updated!" % (lambda_name, version_update, alias_name))
-            return True
-          else:
-            return False
-        except ClientError as error:
-          print(error.response['Error']['Message'])
+        if ARGS.dry_run:
+          print(color.PURPLE + "***Dry run option enabled***" + color.END)
+          print(color.PURPLE + "Would have updated update alias '%s' on version '%s' on lambda '%s'" % (alias_name, version_update, lambda_name) + color.END)
+          return True
+        else:
+          try:
+            update_alias = client.update_alias(
+                            FunctionName='%s' % lambda_name,
+                            Name='%s' % alias_name,
+                            FunctionVersion='%s' % version_update,
+                          )
+            if update_alias['ResponseMetadata']['HTTPStatusCode'] == 200:
+              print("Lamda '%s' version '%s' alias '%s' has been updated!" % (lambda_name, version_update, alias_name))
+              return True
+            else:
+              return False
+          except ClientError as error:
+            print(error.response['Error']['Message'])
       else:
         print("Version not found..")
     else:
